+++ src/stores/NetworkStore/NetworkStore.ts
@@ -12,7 +12,7 @@ export class NetworkStore {
   private shared = createSharedState<SharedState>(
     `network.${this.wallet.instanceId}`,
     {},
-    { readOnly: !this.wallet.isRoot },
+    { readOnly: !this.wallet.isRoot() },
   );
 
   constructor(private wallet: Wallet) {