+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [DAV-2359] Add wallet context support
 - [DAV-3597] Fix authentication issue in Safari
 - [DAV-3136] Show default app banner in case custom banner is not provided
+- [DAV-3126] Add wallet loading state
 
 ### v1.4.0